backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
feedparser==6.0.11
httpx==0.27.0
apscheduler==3.10.4
python-dotenv==1.0.1
cloudscraper==1.2.71
PyJWT==2.8.0
resend==2.29.0
email-validator==2.3.0
curl_cffi==0.15.0
praw==7.8.1
